refactor(skills): unify core skills backend (stage 1)
- move routes/skill.ts to per-entity routes/skill/ folder, tests split per endpoint
- extract toSkillInsertFields + authorizeSkillCreate, dedupe five write surfaces
- extract builtInSkillShippedWrite shared by seed sync and reset
- delete test-only searchSkillCatalogEntries, inline loadFiles
- run_command tool text: "Activated skills" -> "Loaded skills" per glossary
