+++ platform/backend/src/routes/index.ts
@@ -60,7 +60,7 @@ export { default as scheduleTriggerRoutes } from "./schedule-trigger";
 export { default as secretsRoutes } from "./secrets";
 export { default as serviceAccountRoutes } from "./service-account";
 export { default as siteNotificationRoutes } from "./site-notification";
-export { default as skillRoutes } from "./skill";
+export { default as skillRoutes } from "./skill/skill.routes";
 export { default as skillMarketplacePublicRoutes } from "./skill-marketplace-public";
 export { default as skillSandboxArtifactRoutes } from "./skill-sandbox-artifact";
 export { default as skillShareRoutes } from "./skill-share";
