+++ platform/backend/src/routes/skill/catalog-search.skill.route.test.ts
@@ -19,7 +19,7 @@ describe("GET /api/skills/catalog/search", () => {
         organizationId;
     });
 
-    const { default: skillRoutes } = await import("./skill");
+    const { default: skillRoutes } = await import("./skill.routes");
     await app.register(skillRoutes);
   });
