+++ platform/backend/src/routes/skill/list.skill.route.test.ts
@@ -1,0 +1,58 @@
+import type { FastifyInstanceWithZod } from "@/server";
+import { createFastifyInstance } from "@/server";
+import { afterEach, beforeEach, describe, expect, test } from "@/test";
+import type { User } from "@/types";
+
+const MANIFEST = [
+  "---",
+  "name: pdf-processing",
+  "description: Extract text from PDF files.",
+  "---",
+  "",
+  "# PDF Processing",
+  "Use pdftotext -layout.",
+].join("\n");
+
+describe("GET /api/skills", () => {
+  let app: FastifyInstanceWithZod;
+  let user: User;
+  let organizationId: string;
+
+  beforeEach(async ({ makeOrganization, makeUser }) => {
+    user = await makeUser();
+    organizationId = (await makeOrganization()).id;
+
+    app = createFastifyInstance();
+    app.addHook("onRequest", async (request) => {
+      (request as typeof request & { user: unknown }).user = user;
+      (request as typeof request & { organizationId: string }).organizationId =
+        organizationId;
+    });
+
+    const { default: skillRoutes } = await import("./skill.routes");
+    await app.register(skillRoutes);
+  });
+
+  afterEach(async () => {
+    await app.close();
+  });
+
+  test("lists skills with a file count that includes SKILL.md", async () => {
+    await app.inject({
+      method: "POST",
+      url: "/api/skills",
+      payload: {
+        content: MANIFEST,
+        files: [{ path: "references/FORMS.md", content: "# Forms" }],
+      },
+    });
+
+    const response = await app.inject({ method: "GET", url: "/api/skills" });
+
+    expect(response.statusCode).toBe(200);
+    const body = response.json();
+    expect(body.data).toHaveLength(1);
+    // one bundled resource (references/FORMS.md) plus the SKILL.md manifest.
+    expect(body.data[0].fileCount).toBe(2);
+  });
+});
